권한파일업로드

+++ django/comment/articles/views.py
@@ -3,6 +3,7 @@ from django.shortcuts import render, redirect, get_object_or_404
 # django가 주는 views에서 쓸 decorators http를 위한
 from django.views.decorators.http import require_POST
 from django.contrib.auth.decorators import login_required
+from django.contrib import messages
 from .models import Article, Comment
 from .forms import ArticleForm, CommentForm
 from IPython import embed
@@ -31,10 +32,15 @@ def detail(request, article_pk):
 @login_required
 def create(request):
     if request.method == "POST":
-        form = ArticleForm(request.POST)
+        form = ArticleForm(request.POST, request.FILES)
         if form.is_valid():
-            article = form.save()
+            article = form.save(commit=False)
+            article.user = request.user
+            form.save()
+            messages.success(request, '게시글 작성 완료!')
             return redirect('articles:detail', article.pk)
+        else:
+            messages.error(request, '잘못된 데이터입니다.')
     else:
         form = ArticleForm()
     context = {
@@ -46,9 +52,10 @@ def create(request):
 def update(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
     if request.method == "POST":
-        form = ArticleForm(request.POST, instance=article)
+        form = ArticleForm(request.POST, request.FILES, instance=article)
         if form.is_valid():
             article = form.save()
+            messages.info(request, '게시글이 생성되었습니다.')
             return redirect('articles:detail', article.pk)
     else:
         form = ArticleForm(instance=article)
@@ -61,8 +68,12 @@ def update(request, article_pk):
 def delete(request, article_pk):
     if request.user.is_authenticated:
         article = get_object_or_404(Article, pk=article_pk)
-        article.delete()
-        return redirect('articles:index')
+        if article.user == request.user:
+            article.delete()
+            return redirect('articles:index')
+        else:
+            return redirect('articles:detail', article_pk)
+    return redirect('articles:login')
         
 @login_required
 @require_POST
@@ -74,8 +85,8 @@ def comment_create(request, article_pk):
         if comment_form.is_valid():
             comment = comment_form.save(commit=False)
             comment.article = article
-            # comment.article_id = article.pk
-            comment.save()
+            comment.user = request.user
+            comment_form.save()
             return redirect('articles:detail', article.pk)
         else:
             context = {
